+++ src/degasolv/util.clj
@@ -4,10 +4,12 @@
             [clojure.pprint :as pprint]
             [clj-http.client :as client]))
 
-; Deprecated, we should just use update-in
-(defn assoc-conj
-  [mp k v]
-  (update-in mp [k] conj v))
+
+(defmacro dbg [body]
+  `(let [x# ~body]
+     (println "dbg:" '~body "=" x#)
+     (flush)
+     x#))
 
 ; UTF-8 by default :)
 (defn base-slurp [loc]
@@ -17,23 +19,35 @@
     (clojure.core/slurp input :encoding "UTF-8")))
 
 (defn default-slurp [resource]
-  (if-let [[whole-thing protocol auth-stuff rest-of-it]
-           (re-matches #"(https?://)([^@]+)@(.+)" resource)]
-    (:body (if-let [[_ username password]
-             (re-matches #"([^:]+):([^:]+)" auth-stuff)]
-      (client/get (str
-                    protocol
-                    rest-of-it)
-                  {:basic-auth [(java.net.URLDecoder/decode username)
-                                (java.net.URLDecoder/decode password)]})
-      (client/get (str
-                    protocol
-                    rest-of-it)
-                  {:headers
-                   {"Authorization" (str "Bearer "
-                                         (java.net.URLDecoder/decode
-                                           auth-stuff))}})))
-    (base-slurp resource)))
+  (if (re-matches #"https?://.*" resource)
+    (if-let [[whole-thing protocol auth-stuff rest-of-it]
+             (re-matches #"(https?://)([^@]+)@(.+)" resource)]
+      (:body (if-let [[_ username password]
+                      (re-matches #"([^:]+):([^:]+)" auth-stuff)]
+                 (client/get (str
+                                protocol
+                                rest-of-it)
+                   {
+                    :basic-auth [(java.net.URLDecoder/decode username)
+                                 (java.net.URLDecoder/decode password)]})
+               (if-let [[_ headerkey headerval]
+                        (re-matches #"([^=]+)=([^=]+)" auth-stuff)]
+               (client/get (dbg (str
+                             protocol
+                             rest-of-it))
+                           (dbg {
+                            :headers
+                            {
+                             (keyword (java.net.URLDecoder/decode headerkey))
+                             (java.net.URLDecoder/decode headerval)}}))
+               (client/get (str
+                             protocol
+                             rest-of-it)
+                           {
+                            :oauth-token (java.net.URLDecoder/decode auth-stuff)
+                            }))))
+      (client/get resource))
+      (base-slurp resource)))
 
 
 (defn default-spit [loc stuff]